third_party/blink/web_tests/virtual/controls-refresh/month-picker/month-picker-escape-cancellation.html

<!DOCTYPE html>
<html>
<head>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script src='../../../fast/forms/resources/picker-common.js'></script>
</head>
<body>
<input type='month' id='month'>
<script>
'use strict';

promise_test(() => {
  let monthControl = document.getElementById('month');
  return openPickerWithPromise(monthControl)
  .then(() => {
    internals.pagePopupWindow.focus();
    const popupDocument = internals.pagePopupWindow.document;

    eventSender.keyDown('ArrowRight');
    assert_not_equals(internals.pagePopupWindow, null, "Popup should still be open");
    assert_equals(monthControl.value, "", "Arrow key navigation in month popup should not change value");

    eventSender.keyDown('Escape');
    assert_equals(internals.pagePopupWindow, null, "Escape key should close month popup.");
    assert_equals(monthControl.value, "", "Closing month popup with escape key should not commit value");
  });
}, "Month picker: Pressing Escape key closes popup without changing value");

</script>
</body>
</html>
